+++ src/cli/keepassxc-cli.cpp
@@ -16,14 +16,20 @@
  */
 
 #include <cstdlib>
+#include <memory>
 
 #include <QCommandLineParser>
 #include <QCoreApplication>
+#include <QDir>
+#include <QScopedPointer>
 #include <QStringList>
 
 #include "cli/TextStream.h"
 #include <cli/Command.h>
 
+#include "DatabaseCommand.h"
+#include "Open.h"
+#include "Utils.h"
 #include "config-keepassx.h"
 #include "core/Bootstrap.h"
 #include "core/Tools.h"
@@ -32,6 +38,138 @@
 #if defined(WITH_ASAN) && defined(WITH_LSAN)
 #include <sanitizer/lsan_interface.h>
 #endif
+
+#if defined(USE_READLINE)
+#include <readline/history.h>
+#include <readline/readline.h>
+#endif
+
+class LineReader
+{
+public:
+    virtual ~LineReader() = default;
+    virtual QString readLine(QString prompt) = 0;
+    virtual bool isFinished() = 0;
+};
+
+class SimpleLineReader : public LineReader
+{
+public:
+    SimpleLineReader()
+        : inStream(stdin, QIODevice::ReadOnly)
+        , outStream(stdout, QIODevice::WriteOnly)
+        , finished(false)
+    {
+    }
+
+    QString readLine(QString prompt) override
+    {
+        outStream << prompt;
+        outStream.flush();
+        QString result = inStream.readLine();
+        if (result.isNull()) {
+            finished = true;
+        }
+        return result;
+    }
+
+    bool isFinished() override
+    {
+        return finished;
+    }
+
+private:
+    TextStream inStream;
+    TextStream outStream;
+    bool finished;
+};
+
+#if defined(USE_READLINE)
+class ReadlineLineReader : public LineReader
+{
+public:
+    ReadlineLineReader()
+        : finished(false)
+    {
+    }
+
+    QString readLine(QString prompt) override
+    {
+        char* result = readline(prompt.toStdString().c_str());
+        if (!result) {
+            finished = true;
+            return {};
+        }
+        add_history(result);
+        QString qstr(result);
+        free(result);
+        return qstr;
+    }
+
+    bool isFinished() override
+    {
+        return finished;
+    }
+
+private:
+    bool finished;
+};
+#endif
+
+void enterInteractiveMode(const QStringList& arguments)
+{
+    // Replace command list with interactive version
+    Commands::setupCommands(true);
+
+    Open o;
+    QStringList openArgs(arguments);
+    openArgs.removeFirst();
+    o.execute(openArgs);
+
+    QScopedPointer<LineReader> reader;
+#if defined(USE_READLINE)
+    reader.reset(new ReadlineLineReader());
+#else
+    reader.reset(new SimpleLineReader());
+#endif
+
+    QSharedPointer<Database> currentDatabase(o.currentDatabase);
+
+    QString command;
+    while (true) {
+        TextStream errorTextStream(Utils::STDERR, QIODevice::WriteOnly);
+
+        QString prompt;
+        if (currentDatabase) {
+            prompt += currentDatabase->metadata()->name();
+            if (prompt.isEmpty()) {
+                prompt += QFileInfo(currentDatabase->filePath()).fileName();
+            }
+        }
+        prompt += "> ";
+        command = reader->readLine(prompt);
+        if (reader->isFinished()) {
+            return;
+        }
+
+        QStringList args = Utils::splitCommandString(command);
+        if (args.empty()) {
+            continue;
+        }
+
+        auto cmd = Commands::getCommand(args[0]);
+        if (!cmd) {
+            errorTextStream << QObject::tr("Unknown command %1").arg(args[0]) << "\n";
+            continue;
+        } else if (cmd->name == "quit" || cmd->name == "exit") {
+            return;
+        }
+
+        cmd->currentDatabase = currentDatabase;
+        cmd->execute(args);
+        currentDatabase = cmd->currentDatabase;
+    }
+}
 
 int main(int argc, char** argv)
 {
@@ -44,6 +182,7 @@ int main(int argc, char** argv)
     QCoreApplication::setApplicationVersion(KEEPASSXC_VERSION);
 
     Bootstrap::bootstrap();
+    Commands::setupCommands(false);
 
     TextStream out(stdout);
     QStringList arguments;
@@ -54,7 +193,7 @@ int main(int argc, char** argv)
 
     QString description("KeePassXC command line interface.");
     description = description.append(QObject::tr("\n\nAvailable commands:\n"));
-    for (Command* command : Command::getCommands()) {
+    for (auto& command : Commands::getCommands()) {
         description = description.append(command->getDescriptionLine());
     }
     parser.setApplicationDescription(description);
@@ -84,9 +223,13 @@ int main(int argc, char** argv)
     }
 
     QString commandName = parser.positionalArguments().at(0);
-    Command* command = Command::getCommand(commandName);
+    if (commandName == "open") {
+        enterInteractiveMode(arguments);
+        return EXIT_SUCCESS;
+    }
 
-    if (command == nullptr) {
+    auto command = Commands::getCommand(commandName);
+    if (!command) {
         qCritical("Invalid command %s.", qPrintable(commandName));
         // showHelp exits the application immediately, so we need to set the
         // exit code here.